@@ -5,9 +5,10 @@
     }
   },
   "require": {
-    "php": "^7.2",
+    "php": ">=8.0",
     "ext-json": "*",
     "predis/predis": "^1.1",
-    "guzzlehttp/guzzle": "^6.3"
+    "guzzlehttp/guzzle": "^6.3",
+    "sentry/sdk": "^3.1"
   }
 }
