@@ -7,13 +7,15 @@ use Predis\Client as PredisClient;
 use Rocket\{Response, Validator, Api};
 
 require __DIR__ . '/../vendor/autoload.php';
+$config = include_once __DIR__ . '/../config/parameters.php';
+
+Sentry\init(['dsn' => $config['sentry_key']]);
 
 header("Content-Type: application/json");
 
 $response = new Response();
 
 try {
-
     $validator = new Validator($_GET);
 
     $validator->check();
